% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_datasets.R
\name{as_dataloader.list}
\alias{as_dataloader.list}
\title{as_dataloader.list}
\usage{
\method{as_dataloader}{list}(x, ...)
}
\arguments{
\item{x}{the list to convert}
}
\description{
Override generic function to handle list(x=DelayedArray,y=factor).
}
