% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_sampling_dataset.R
\name{bag_sampling_classification_dataset}
\alias{bag_sampling_classification_dataset}
\title{bag_sampling_classification_dataset}
\usage{
bag_sampling_classification_dataset(
  x,
  sample_ids,
  y,
  nbag = 1000L,
  bag_size = 100L,
  seed = 1234L,
  replace = TRUE,
  post_process = identity,
  balanced = TRUE
)
}
\arguments{
\item{h5f}{a character vector of h5 filenames (recycled to maximum length of the arguments)}

\item{name}{a character vector of array names in the h5 file (recycled to maximum length of the arguments)}

\item{perm}{a list of integer vectors of array permuations to apply (recylced if needed).}

\item{index}{a list of `alist()` for array subsetting}

\item{transform}{a unary function to be apply on each element of the dataset}

\item{transform_h5elt}{a unary function to be apply on the list of raw HDF5Array (e.g. permutation and subset indexing)}
}
\description{
Given labelled bags of instances, randomly generate new bags bag_size by random sampling random sample instances
}
\examples{
x <- Matrix(0,3000,34)
sample_ids <- gl(3,1000)
y <- c("a","b","a")
bag_sampling_classification_dataset(x,sample_ids,y,nbag=10,bag_size=100)[1:3]
}
