% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_datasets.R
\name{delayed_classifier_dataset}
\alias{delayed_classifier_dataset}
\title{delayed_classifier_dataset}
\usage{
delayed_classifier_dataset(
  x,
  y = NULL,
  x_dtype = torch_float(),
  y_dtype = torch_long()
)
}
\arguments{
\item{x}{Expression matrix.}

\item{y}{Cell type array.}

\item{x_dtype}{torch types.}

\item{y_dtype}{torch types.}
}
\description{
A torch dataset which accepts DelayedArrays.
}
