% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luz_ext.R
\name{luz_callback_lr_geometric}
\alias{luz_callback_lr_geometric}
\title{luz_callback_lr_geometric}
\usage{
luz_callback_lr_geometric(epochs, lr_start = 0.001, lr_end = 1e-06)
}
\arguments{
\item{epochs}{a vector of positive integers, defining a consecutive number of
epochs.}

\item{lr_start}{a vector of initial learning rate which will be recycled to
length(epochs)}

\item{lr_end}{a vector of ending learning rate which will be recycled to
length(epochs).}
}
\description{
luz_callback_lr_geometric
}
