% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luz_ext.R
\name{luz_callback_prune}
\alias{luz_callback_prune}
\title{luz_callback_prune}
\usage{
luz_callback_prune(param, n, at_epoch)
}
\arguments{
\item{param}{Model parameters}

\item{n}{number of top/bottom features to keep after pruning}

\item{at_epoch}{Epoch at which compute pruning mask.}
}
\description{
luz_callback_prune
}
