% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luz_ext.R
\name{luz_metric_hinge_accuracy}
\alias{luz_metric_hinge_accuracy}
\title{luz_metric_hinge_accuracy}
\usage{
luz_metric_hinge_accuracy()
}
\description{
Accuracy metric for 1d outputs and numeric hinge targets (-1,+1)
}
