% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_module.R
\name{nn_bag_softmax_pool1d}
\alias{nn_bag_softmax_pool1d}
\title{nn_bag_softmax_pool1d}
\usage{
nn_bag_softmax_pool1d(in_features, out_features)
}
\arguments{
\item{in_features}{number of input dimensions}

\item{out_features}{number of output dimensions}
}
\description{
Pool bag of elements with softmax weights
}
\examples{
nn_bag_softmax_pool1d(32L,2L)(torch_rand(3L,5L,32L))
}
