% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_module.R
\name{nn_cell_scorer}
\alias{nn_cell_scorer}
\title{nn_cell_scorer}
\usage{
nn_cell_scorer(
  feature_names,
  class_names,
  input_dropout_rate = 0.5,
  n = 256L,
  dropout_rates = 0.25,
  ...
)
}
\arguments{
\item{feature_names}{Names of the features.}

\item{class_names}{Class of the cells.}

\item{input_dropout_rate}{Dropout rate to apply on input values when training}

\item{n}{A numeric integer vector of length>0 specifying the number of feature in each internal layer.}

\item{dropout_rates}{A numeric vector of length>0 of dropouts to apply before each internal layer.}

\item{...}{additional parameters are passed to nn_linear_rescale() when creating first layer.}
}
\description{
Cell type classifier NN-based.
}
\examples{
nn_cell_scorer(1:4,1:3)(torch::torch_rand(32,4))
}
