% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_loss.R
\name{nn_hinge_loss}
\alias{nn_hinge_loss}
\title{nn_hinge_loss}
\usage{
nn_hinge_loss(margin = 1, reduction = "mean")
}
\arguments{
\item{margin}{a scalar positive value.}

\item{reduction}{name of the reduction to apply}
}
\description{
Hinge regression loss module computing
}
\examples{
nn_hinge_loss(reduction="none")(torch_tensor(c(-2,-0.5,+0.7,3)),torch_tensor(c(-1,-1,+1,+1)))
}
