% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_loss.R
\name{nn_ordinal_regression_loss}
\alias{nn_ordinal_regression_loss}
\title{nn_ordinal_regression_loss}
\usage{
nn_ordinal_regression_loss(C)
}
\arguments{
\item{C}{a square 2D float tensor representing the cost matrix.}
}
\description{
Ordinal regression loss module
}
\examples{
nn_ordinal_regression_loss(1 - torch::torch_eye(3L))
}
