% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_loss.R
\name{nnf_dag_loss}
\alias{nnf_dag_loss}
\title{nnf_dag_loss}
\usage{
nnf_dag_loss(pred, target, e, w, margin = 1)
}
\arguments{
\item{pred}{Predicted values.}

\item{target}{Target values.}

\item{e}{a 2 column integer matrix (edge list of the DAG), with parent index
in the first column and child index in the second}

\item{w}{weight vector for each node of the dag (how much to penalize a node)}

\item{margin}{a scalar positive value.}
}
\value{
loss value
}
\description{
Direct Acyclic Graph loss function.
}
