% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_loss.R
\name{nnf_ordinal_regression_loss}
\alias{nnf_ordinal_regression_loss}
\title{nnf_ordinal_regression_loss}
\usage{
nnf_ordinal_regression_loss(pred, target, C, margin = 1)
}
\arguments{
\item{pred}{Predicted values.}

\item{target}{Target values.}

\item{C}{a square 2D float tensor representing the cost matrix.}

\item{margin}{a scalar positive value.}
}
\value{
loss value
}
\description{
Ordinal regression loss function.
}
