% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delayed.R
\name{predict_delayed}
\alias{predict_delayed}
\title{predict_delayed}
\usage{
predict_delayed(N, luz_model, ignore.case = FALSE, batch_size = 1024L, ...)
}
\arguments{
\item{N}{expression matrix, ideally log1p(RPM/100)/log(2) counts, but raw counts seems to work too.}

\item{luz_model}{fitted model to use for predictions}

\item{ignore.case}{if TRUE ignore case when aligning feature names (=gene names). It is useful for easy Human-Mouse gene-name matching.}

\item{batch_size}{Size of data loaded at training.}

\item{...}{Additional argument to be passed to predict function.}
}
\value{
A prediction matrix giving each input cell a score for each cell type.
}
\description{
Predict cell type given a torch model.
}
