% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delayed.R
\name{train_delayed_classifier}
\alias{train_delayed_classifier}
\title{train_delayed_classifier}
\usage{
train_delayed_classifier(
  x,
  y,
  pre_pruning_epoch = 25L,
  post_pruning_epoch = 25L,
  batch_size = 256L,
  prune_size = pmin(50L, ncol(x)),
  lr_start = 0.001,
  lr_end = 1e-05,
  attention = TRUE,
  initial_weights_path = NA,
  autoresume_path = NA,
  n = c(4096L, 256L),
  dropout_rates = 0.25,
  weight_decay = 0,
  input_dropout_rate = 0.5,
  y_type = c("nominal", "ordinal"),
  C = NULL,
  callbacks = NULL,
  ...
)
}
\arguments{
\item{x}{Count matrix.}

\item{y}{Array containing the cell type class label.}

\item{pre_pruning_epoch}{Number of epoch before pruning}

\item{post_pruning_epoch}{Number of epoch after pruning. Set to 0}

\item{batch_size}{Size of batch of data loaded at each training step.}

\item{prune_size}{number of top/down gene to keep for each node of layer1 when pruning}

\item{lr_start}{initial learning rate (both at epoch 1 and after pruning)}

\item{lr_end}{final learning rate (both at pruning and at last epoch)}

\item{attention}{wether to use attention in the first layer (cosine_similarity layer)}

\item{initial_weights_path}{a character to specify a file to load initial weights from.
Set to NA to disable.}

\item{autoresume_path}{a character to specify a file to store model weights at each epoch.
Set to NA to disable.}

\item{n}{A numeric integer vector of length>0 specifying the number of feature in each internal layer.}

\item{dropout_rates}{A numeric vector of length>0 of dropouts to apply before each internal layer.}

\item{weight_decay}{value passed to the optim_adam optimizer.}

\item{input_dropout_rate}{Dropout rate to apply on input values when training}

\item{y_type}{type of target factor: either nominal or ordinal (default: nominal)}

\item{C}{the cost matrix to use in ordinal regression}

\item{callbacks}{a list of additional callbacks to use}

\item{...}{additional parameters are passed to fit() call}
}
\description{
Allows user-friendly training set up of a torch model.
}
