% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_dataset_bags.R
\name{bag_sampling_dataset}
\alias{bag_sampling_dataset}
\title{bag_sampling_dataset}
\usage{
bag_sampling_dataset(
  x,
  bags,
  y,
  nbag = 1000L,
  bag_size = 100L,
  seed = 1234L,
  replace = TRUE,
  replace_elt = replace,
  post_process = identity,
  balancing = TRUE
)
}
\arguments{
\item{x}{a generic 2D matrix object, with elements in rows and features as column}

\item{bags}{a list of integer vectors referencing rows of x}

\item{y}{a vector of target values for each bag, so the length must match nlevels(sample_ids)}

\item{nbag}{number of random bags to generate in the dataset}

\item{bag_size}{size of the bags to generate}

\item{seed}{set seed for random sampling}

\item{replace}{a logical, if TRUE the same sample might be sampled multiple time.}

\item{replace_elt}{a logical, if TRUE the same elements might be sampled multiple time.}

\item{post_process}{a function to use to post_process the batches}

\item{balancing}{a grouping factor for balancing}
}
\description{
Given labelled bags of instances, randomly generate new bags bag_size by random sampling random sample instances
}
\examples{
bag_sampling_dataset(
  matrix((1:1000-1)\%\%100+1,100,10),
  split(seq(100),gl(20,5)),
  gl(2,10),
  nbag = 7L,bag_size=5L,balanced=TRUE
)[1:3]
}
