% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_datasets.R
\name{h5_dataset}
\alias{h5_dataset}
\title{delayed_classifier_dataset}
\usage{
h5_dataset(
  x,
  y = NULL,
  name.x = NULL,
  name.y = NULL,
  row_idx = NULL,
  col_idx.x = NULL
)
}
\arguments{
\item{x}{Expression matrix.}

\item{y}{Cell type array.}
}
\description{
A torch dataset which accepts DelayedArrays.
}
\examples{
#x <- h5_dataset("../../data/ABAMouseTENx/sce/assays.h5",name.x="/assay001*")
}
