% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_loss.R
\name{nn_dag_loss}
\alias{nn_dag_loss}
\title{nn_dag_loss}
\usage{
nn_dag_loss(dag, w = 1)
}
\arguments{
\item{dag}{a 2 column integer matrix (edge list of the DAG), with parent index
in the first column and child index in the second}

\item{w}{(m*n) float tensor of weights for each sample and node in the dag}
}
\description{
Direct Acyclic Graph loss module
}
