% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_loss.R
\name{nnf_dag_loss}
\alias{nnf_dag_loss}
\title{nnf_dag_loss}
\usage{
nnf_dag_loss(pred, target, e, w, margin = 1)
}
\arguments{
\item{pred}{(m*n) float tensor of predicted values, where m is number of
sample in the batch and n is number of node in the dag.}

\item{target}{a (m*n) binary tensor of target classes.}

\item{e}{a 2 column integer matrix (edge list of the DAG), with parent index
in the first column and child index in the second.}

\item{w}{float tensor of weights compatible with (m*n) tensor (how much to
penalize a sample/node)}

\item{margin}{a scalar positive value.}
}
\value{
loss value
}
\description{
Direct Acyclic Graph loss function.
}
