% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_loss.R
\name{nnf_hinge_loss}
\alias{nnf_hinge_loss}
\title{nnf_hinge_loss}
\usage{
nnf_hinge_loss(input, target, margin = 1, reduction = c("mean", "sum", "none"))
}
\arguments{
\item{target}{Binary target value {-1,+1}}

\item{margin}{a scalar positive value.}

\item{pred}{Predicted values.}
}
\value{
hinge loss scalar value
}
\description{
Hinge loss
}
